Funny fix of a disaster (possibly stuff missing, but it runs)
So I messed up and made edits that weren’t in git as commits. Grr. What
happened was I misspelled “missile” as “missle” but wasn’t sure how to
do a case sensitive search in the editor so I couldn’t fix it in any
good way. Will try to fix now.